Added soft mask creation to utilities; removed soft masking from ISAC since it's not actually useful here
